PDDA-409 Fixing pmd issues following v7.5 update

+++ src/main/java/uk/gov/hmcts/pdda/business/services/cppformatting/CppFormattingControllerBean.java
@@ -147,7 +147,7 @@ public class CppFormattingControllerBean extends AbstractControllerBean implemen
      */
     public void refreshPublicDisplaysForCourt(Integer courtId) {
         LOG.debug("refreshPublicDisplaysForCourt({})", courtId);
-        CourtConfigurationChange ccc = new CourtConfigurationChange(courtId.intValue(), true);
+        CourtConfigurationChange ccc = new CourtConfigurationChange(courtId, true);
         ConfigurationChangeEvent ccEvent = new ConfigurationChangeEvent(ccc);
         getPublicDisplayNotifier().sendMessage(ccEvent);
     }
